added husky + pretty-quick + moved folders
