+++ public/index.html
@@ -8,7 +8,7 @@
       name="description"
       content="Web site created using create-react-app"
     />
-    <link rel="stylesheet" href="https://use.typekit.net/dcx5exd.css">
+    <link rel="stylesheet" href="https://use.typekit.net/dcx5exd.css" />
     <!--
       manifest.json provides metadata used when your web app is installed on a
       user's mobile device or desktop. See https://developers.google.com/web/fundamentals/web-app-manifest/
