--- spotif-i/src/App.tsx
@@ -1,14 +1,0 @@
-import React from "react";
-import { HomePage } from "./pages/HomePage";
-
-function App() {
-    return (
-        <div className="App">
-            <header className="App-header">
-                <HomePage />
-            </header>
-        </div>
-    );
-}
-
-export default App;
