--- spotif-i/src/components/svg/Bird.tsx
@@ -1,61 +1,0 @@
-import React from "react";
-
-export const Bird = () => {
-    return (
-        <svg
-            width="500"
-            height="300"
-            viewBox="0 0 712 380"
-            fill="none"
-            xmlns="http://www.w3.org/2000/svg"
-        >
-            <path
-                d="M591.229 320.299C549.535 306.175 511.018 283.596 478.754 254.292C468.529 245.069 458.999 235.172 448.476 226.237C427.729 208.558 403.705 194.915 379.086 182.713C352.382 169.55 324.288 157.348 303.739 136.307C290.337 122.471 280.906 105.561 270.681 89.4198C260.456 73.2785 248.742 57.3293 232.462 46.8566C216.082 36.4801 193.944 32.6369 176.77 41.7644C153.144 54.3508 147.684 85.8649 154.037 111.23C160.291 136.691 175.182 159.462 182.329 184.635C188.186 205.484 188.584 227.39 190.172 249.008C191.959 273.124 195.433 297.528 205.559 319.722C215.685 341.821 233.156 361.709 256.485 370.548C276.935 378.331 299.768 376.986 321.707 375.545C340.767 374.296 359.827 373.046 384.149 369.011C488.979 386.113 597.88 379.196 699.435 348.931C637.39 351.717 575.048 347.97 513.896 337.881C566.411 332.981 620.316 343.358 671.639 331.828C643.545 332.597 615.451 328.562 591.229 320.299Z"
-                stroke="black"
-                stroke-width="3"
-                stroke-miterlimit="10"
-            />
-            <path
-                d="M174.586 107.579C161.482 106.618 148.379 104.408 135.672 100.949C142.919 99.316 150.165 98.163 157.512 97.3944C142.025 93.1669 127.631 85.7688 113.435 78.4667C114.328 77.0255 116.314 76.6412 118.001 76.4491C128.226 75.0079 138.65 73.7589 148.875 75.2C159.001 76.4491 169.226 80.9648 174.586 107.579Z"
-                stroke="black"
-                stroke-width="2"
-                stroke-miterlimit="10"
-            />
-            <path
-                d="M201.092 80.9649C200.893 78.1786 203.276 75.3923 207.147 73.9511C208.835 73.2785 210.92 72.8942 212.905 73.1825C214.593 73.4707 215.883 74.1433 216.975 74.9119C219.655 76.8335 221.045 79.3315 220.847 81.7335"
-                stroke="black"
-                stroke-width="2"
-                stroke-miterlimit="10"
-            />
-            <g opacity="0.9804">
-                <g opacity="0.9804">
-                    <path
-                        opacity="0.9804"
-                        d="M382.759 368.915C347.319 353.734 314.361 332.309 288.451 304.446C262.541 276.487 243.779 241.898 237.822 204.715C265.023 250.833 310.489 284.173 356.849 312.612C350.297 305.983 345.234 297.912 342.157 289.265C350.495 291.667 357.345 297.528 363.897 303.101C401.223 335.095 442.222 365.648 490.468 377.85"
-                        stroke="black"
-                        stroke-width="2"
-                        stroke-miterlimit="10"
-                    />
-                </g>
-            </g>
-            <path
-                d="M39.5851 2L61.6607 24.0756"
-                stroke="black"
-                stroke-width="3"
-                stroke-linecap="round"
-            />
-            <path
-                d="M2 97.5328H51.5712"
-                stroke="black"
-                stroke-width="3"
-                stroke-linecap="round"
-            />
-            <path
-                d="M86.843 162.357L105.909 143.291"
-                stroke="black"
-                stroke-width="3"
-                stroke-linecap="round"
-            />
-        </svg>
-    );
-};
