+++ src/components/ListEntry.tsx
@@ -1,0 +1,1 @@
+export const ListEntry = () => {};
