+++ src/pages/StatsPage.tsx
@@ -1,0 +1,1 @@
+export const StatsPage = () => {};
