--- styles/style.css
@@ -1,204 +1,0 @@
-body {
-    background-color: white;
-    overflow-x: hidden;
-}
-
-@font-face {
-    font-family: Bluu;
-    src: url("/fonts/BluuNext-Bold.otf");
-}
-
-#login-button {
-    display: flex;
-    justify-content: center;
-    align-items: center;
-}
-
-.spotify {
-    padding: 10px;
-    padding-right: 15px;
-    border-radius: 25px;
-    border: 0px;
-    font-size: 11pt;
-    box-shadow: 0 1px 5px rgba(0, 0, 0, 0.3);
-    background-color: black;
-    font-family: Helvetica Neue;
-    display: flex;
-    align-items: center;
-    color: white;
-}
-
-#sub-header {
-    text-align: center;
-    font-size: 16pt;
-    font-family: Work Sans;
-    margin-bottom: 20px;
-}
-
-@media screen and (max-width: 700px) {
-    #sub-header {
-        font-size: 12pt;
-    }
-}
-
-.spotify:hover {
-    cursor: pointer;
-}
-
-#icon {
-    margin-right: 10px;
-}
-
-.text {
-    font-family: IBM Plex Mono;
-}
-
-.header {
-    margin-bottom: 20px;
-    font-size: 60pt;
-    margin-top: 40px;
-    text-align: center;
-    font-family: Bluu;
-    color: #a19ffc;
-}
-
-@media screen and (max-width: 700px) {
-    .header {
-        font-size: 40pt;
-    }
-}
-
-.info {
-    text-align: center;
-    margin-bottom: 20px;
-    font-family: Work Sans;
-}
-
-.flex-container {
-    display: flex;
-    align-items: center;
-    flex-direction: column;
-}
-
-#mediumterm-artist,
-#shortterm-artist,
-#longterm-artist,
-#mediumterm,
-#shortterm,
-#longterm {
-    width: 750px;
-    background-color: white;
-    padding: 0 30px 30px 30px;
-    margin-left: 5%;
-}
-
-@media screen and (max-width: 1000px) {
-    #mediumterm-artist,
-    #shortterm-artist,
-    #longterm-artist,
-    #longterm,
-    #mediumterm,
-    #shortterm {
-        width: 600px;
-        background-color: white;
-        padding: 0 15px 15px 15px;
-        font-size: 10pt;
-    }
-}
-
-#topartists-longterm li,
-#toptracks-longterm li,
-#topartists-mediumterm li,
-#toptracks-mediumterm li,
-#topartists-shortterm li,
-#toptracks-shortterm li {
-    display: flex;
-    align-items: center;
-}
-
-#results {
-    margin-bottom: 60px;
-    font-family: "Rubik";
-    font-size: 13pt;
-}
-
-.artist-photo {
-    width: 100px;
-    height: 100px;
-    margin-right: 20px;
-    border-radius: 10px;
-}
-
-.artist-li {
-    margin-top: 10px;
-}
-
-.number {
-    font-family: "Rubik";
-    font-size: 14pt;
-    color: #aea3bf;
-}
-
-#terms {
-    padding-bottom: 40px;
-    font-family: "Rubik";
-    font-weight: 500;
-    font-size: 16pt;
-    display: flex;
-    justify-content: center;
-}
-
-#terms-inner {
-    display: flex;
-}
-
-#terms-inner div {
-    margin-right: 20px;
-}
-
-#footer {
-    position: fixed;
-    top: 45%;
-    height: 10%;
-    left: 0;
-    width: 150px;
-    display: flex;
-    flex-direction: column;
-    justify-content: space-evenly;
-    align-items: center;
-    font-family: Work Sans;
-    margin-left: 1%;
-}
-
-.footer-child {
-    margin-bottom: 10px;
-}
-
-@media screen and (max-width: 1000px) {
-    #footer {
-        margin-left: 0;
-        padding-top: 10px;
-        padding-bottom: 10px;
-        font-family: Work Sans;
-        display: flex;
-        flex-direction: row;
-        box-shadow: 0 1px 5px rgba(0, 0, 0, 0.3);
-        justify-content: space-evenly;
-        position: fixed;
-        height: 40px;
-        bottom: 0;
-        top: unset;
-        width: 100%;
-        background-color: #fff;
-        left: 0;
-        font-size: 10pt;
-        color: #514cec;
-    }
-
-    .footer-child {
-        display: flex;
-        flex-direction: column;
-        justify-content: space-between;
-        align-items: center;
-    }
-}
